Patch for IceConnection Failure during streaming
If an IceConnection Failure occurs during streaming,
the PeerConnection is still active even though the streaming can not be performed.
In this situation, the PeerConnection is in the active state and
can not be streamed internally.
To eliminate this case, if IceConnection Failure occurs,
delete the currently active PeerConnection and make sure that
there is no problem in the next PeerConnection request.