Modify DirectSocket to handle multiple messages #76
In Direct Socket, when multiple peer messages are read from socket at
the same time, only the first message is processed and the remaining
messages are ignored.

This patch modifies DirectSocket to distinguish and process each message
with a DirectSocket delimiter when multiple DirectSocket peer messages
are read at a time.