Add web frontend support for '.local' domains
On Mac OS, by default, it's possible to reach a RasPi on your local network at its `hostname.local` domain, ie. `http://mypihostname.local:8889/np2/`. When I try to load the RWS control panel this way, the webpage loads but it fails the `isPrivateIPaddress` check when initiating a websocket connection. This change updates the `isPrivateIPaddress` function to return true if the host domain ends in `.local`.