Added config for camera roi setting #113
Add video_roi entry to media_config.conf file
The roi setting should be set as x,y,width,height as follows.

video_roi=0.25,0.25,0.25,0.25

Each value is set to a value normalized to a value of 0 ~ 1.0
for the size of the video resolution. The width value should be
set to 0.998 as max. If not in use, delete the video_roi setting value
and use it as the default value.

When the roi is set, initial streaming starts with the video roi,
and all zoom related in/out/move also operate together.